+++ lib/src/pipeline.dart
@@ -117,8 +117,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/bitcount
-  Pipeline bitcount(String key,
-      {int? start, int? end, CommandOption<int, int>? opts}) {
+  Pipeline bitcount(String key, {int? start, int? end, CommandOption<int, int>? opts}) {
     return _chain(BitCountCommand(key, start: start, end: end, opts: opts));
   }
 
@@ -133,8 +132,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/bitpos
-  Pipeline bitpos(String key, int bit,
-      [int? start, int? end, CommandOption<int, int>? opts]) {
+  Pipeline bitpos(String key, int bit, [int? start, int? end, CommandOption<int, int>? opts]) {
     return _chain(BitPosCommand(key, bit, start, end, opts));
   }
 
@@ -224,26 +222,22 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/getrange
-  Pipeline getrange(String key, int start, int end,
-      [CommandOption<String, String>? opts]) {
+  Pipeline getrange(String key, int start, int end, [CommandOption<String, String>? opts]) {
     return _chain(GetRangeCommand(key, start, end, opts));
   }
 
   /// @see https://redis.io/commands/getset
-  Pipeline getset<TData>(String key, TData value,
-      [CommandOption<dynamic, TData>? opts]) {
+  Pipeline getset<TData>(String key, TData value, [CommandOption<dynamic, TData>? opts]) {
     return _chain(GetSetCommand<TData>(key, value, opts));
   }
 
   /// @see https://redis.io/commands/hget
-  Pipeline hget<TData>(String key, String field,
-      [CommandOption<dynamic, TData?>? opts]) {
+  Pipeline hget<TData>(String key, String field, [CommandOption<dynamic, TData?>? opts]) {
     return _chain(HGetCommand<TData>(key, field, opts));
   }
 
   /// @see https://redis.io/commands/hgetall
-  Pipeline hgetall<TData>(String key,
-      [CommandOption<dynamic, Map<String, TData?>>? opts]) {
+  Pipeline hgetall<TData>(String key, [CommandOption<dynamic, Map<String, TData?>>? opts]) {
     return _chain(HGetAllCommand<TData>(key, opts));
   }
 
@@ -268,8 +262,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/hkeys
-  Pipeline hkeys(String key,
-      [CommandOption<List<String>, List<String>>? opts]) {
+  Pipeline hkeys(String key, [CommandOption<List<String>, List<String>>? opts]) {
     return _chain(HKeysCommand(key, opts));
   }
 
@@ -288,8 +281,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/hmset
-  Pipeline hmset<TData>(String key, Map<String, TData> kv,
-      [CommandOption<String, String>? opts]) {
+  Pipeline hmset<TData>(String key, Map<String, TData> kv, [CommandOption<String, String>? opts]) {
     return _chain(HMSetCommand<TData>(key, kv, opts));
   }
 
@@ -321,8 +313,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/hset
-  Pipeline hset<TData>(String key, Map<String, TData> kv,
-      [CommandOption<int, int>? opts]) {
+  Pipeline hset<TData>(String key, Map<String, TData> kv, [CommandOption<int, int>? opts]) {
     return _chain(HSetCommand<TData>(key, kv, opts));
   }
 
@@ -342,8 +333,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/hsetnx
-  Pipeline hsetnx<TData>(String key, String field, TData value,
-      [CommandOption<int, int>? opts]) {
+  Pipeline hsetnx<TData>(String key, String field, TData value, [CommandOption<int, int>? opts]) {
     return _chain(HSetNXCommand<TData>(key, field, value, opts));
   }
 
@@ -353,8 +343,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/hvals
-  Pipeline hvals<TData>(String key,
-      [CommandOption<List<TData>, List<TData>>? opts]) {
+  Pipeline hvals<TData>(String key, [CommandOption<List<TData>, List<TData>>? opts]) {
     return _chain(HValsCommand<TData>(key, opts));
   }
 
@@ -374,14 +363,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/keys
-  Pipeline keys(String pattern,
-      [CommandOption<List<String>, List<String>>? opts]) {
+  Pipeline keys(String pattern, [CommandOption<List<String>, List<String>>? opts]) {
     return _chain(KeysCommand(pattern, opts));
   }
 
   /// @see https://redis.io/commands/lindex
-  Pipeline lindex<TData>(String key, int index,
-      [CommandOption<Object?, TData?>? opts]) {
+  Pipeline lindex<TData>(String key, int index, [CommandOption<Object?, TData?>? opts]) {
     return _chain(LIndexCommand<TData>(key, index, opts));
   }
 
@@ -399,6 +386,22 @@ class Pipeline {
   /// @see https://redis.io/commands/llen
   Pipeline llen(String key, [CommandOption<int, int>? opts]) {
     return _chain(LLenCommand(key, opts));
+  }
+
+  /// @see https://redis.io/commands/lmove
+  Pipeline lmove<TData>(
+    String source,
+    String destination, {
+    required LMoveDir whereFrom,
+    required LMoveDir whereTo,
+    CommandOption<TData, TData>? opts,
+  }) {
+    return _chain(LMoveCommand<TData>(
+      source,
+      destination,
+      whereFrom: whereFrom,
+      whereTo: whereTo,
+    ));
   }
 
   /// @see https://redis.io/commands/lpop
@@ -426,14 +429,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/lpush
-  Pipeline lpush<TData>(String key, List<TData> elements,
-      [CommandOption<int, int>? opts]) {
+  Pipeline lpush<TData>(String key, List<TData> elements, [CommandOption<int, int>? opts]) {
     return _chain(LPushCommand<TData>(key, elements, opts));
   }
 
   /// @see https://redis.io/commands/lpushx
-  Pipeline lpushx<TData>(String key, List<TData> elements,
-      [CommandOption<int, int>? opts]) {
+  Pipeline lpushx<TData>(String key, List<TData> elements, [CommandOption<int, int>? opts]) {
     return _chain(LPushCommand<TData>(key, elements, opts));
   }
 
@@ -448,8 +449,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/lrem
-  Pipeline lrem<TData>(String key, int count, TData value,
-      [CommandOption<int, int>? opts]) {
+  Pipeline lrem<TData>(String key, int count, TData value, [CommandOption<int, int>? opts]) {
     return _chain(LRemCommand<TData>(key, count, value, opts));
   }
 
@@ -482,14 +482,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/mset
-  Pipeline mset<TData>(Map<String, TData> kv,
-      [CommandOption<String, String>? opts]) {
+  Pipeline mset<TData>(Map<String, TData> kv, [CommandOption<String, String>? opts]) {
     return _chain(MSetCommand<TData>(kv, opts));
   }
 
   /// @see https://redis.io/commands/msetnx
-  Pipeline msetnx<TData>(Map<String, TData> kv,
-      [CommandOption<int, int>? opts]) {
+  Pipeline msetnx<TData>(Map<String, TData> kv, [CommandOption<int, int>? opts]) {
     return _chain(MSetNXCommand<TData>(kv, opts));
   }
 
@@ -499,14 +497,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/pexpire
-  Pipeline pexpire(String key, int milliseconds,
-      [CommandOption<dynamic, int>? opts]) {
+  Pipeline pexpire(String key, int milliseconds, [CommandOption<dynamic, int>? opts]) {
     return _chain(PExpireCommand(key, milliseconds, opts));
   }
 
   /// @see https://redis.io/commands/pexpireat
-  Pipeline pexpireat(String key, int unix,
-      [CommandOption<dynamic, int>? opts]) {
+  Pipeline pexpireat(String key, int unix, [CommandOption<dynamic, int>? opts]) {
     return _chain(PExpireAtCommand(key, unix, opts));
   }
 
@@ -531,8 +527,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/publish
-  Pipeline publish<TData>(String channel, TData message,
-      [CommandOption<int, int>? opts]) {
+  Pipeline publish<TData>(String channel, TData message, [CommandOption<int, int>? opts]) {
     return _chain(PublishCommand<TData>(channel, message, opts));
   }
 
@@ -542,14 +537,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/rename
-  Pipeline rename(String source, String destination,
-      [CommandOption<String, String>? opts]) {
+  Pipeline rename(String source, String destination, [CommandOption<String, String>? opts]) {
     return _chain(RenameCommand(source, destination, opts));
   }
 
   /// @see https://redis.io/commands/renamenx
-  Pipeline renamenx(String source, String destination,
-      [CommandOption<dynamic, int>? opts]) {
+  Pipeline renamenx(String source, String destination, [CommandOption<dynamic, int>? opts]) {
     return _chain(RenameNXCommand(source, destination, opts));
   }
 
@@ -559,20 +552,17 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/rpush
-  Pipeline rpush<TData>(String key, List<TData> elements,
-      [CommandOption<int, int>? opts]) {
+  Pipeline rpush<TData>(String key, List<TData> elements, [CommandOption<int, int>? opts]) {
     return _chain(RPushCommand<TData>(key, elements, opts));
   }
 
   /// @see https://redis.io/commands/rpushx
-  Pipeline rpushx<TData>(String key, List<TData> elements,
-      [CommandOption<int, int>? opts]) {
+  Pipeline rpushx<TData>(String key, List<TData> elements, [CommandOption<int, int>? opts]) {
     return _chain(RPushXCommand<TData>(key, elements, opts));
   }
 
   /// @see https://redis.io/commands/sadd
-  Pipeline sadd<TData>(String key, List<TData> members,
-      [CommandOption<int, int>? opts]) {
+  Pipeline sadd<TData>(String key, List<TData> members, [CommandOption<int, int>? opts]) {
     return _chain(SAddCommand<TData>(key, members, opts));
   }
 
@@ -584,8 +574,7 @@ class Pipeline {
     String? type,
     CommandOption<List<dynamic>, List<dynamic>>? opts,
   }) {
-    return _chain(ScanCommand(cursor,
-        match: match, count: count, opts: opts, type: type));
+    return _chain(ScanCommand(cursor, match: match, count: count, opts: opts, type: type));
   }
 
   /// @see https://redis.io/commands/scard
@@ -594,14 +583,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/script-exists
-  Pipeline scriptExists(List<String> hashes,
-      [CommandOption<List<int>, List<int>>? opts]) {
+  Pipeline scriptExists(List<String> hashes, [CommandOption<List<int>, List<int>>? opts]) {
     return _chain(ScriptExistsCommand(hashes, opts));
   }
 
   /// @see https://redis.io/commands/script-flush
-  Pipeline scriptFlush(
-      {bool? sync, bool? async, CommandOption<String, String>? opts}) {
+  Pipeline scriptFlush({bool? sync, bool? async, CommandOption<String, String>? opts}) {
     return _chain(ScriptFlushCommand(sync: sync, async: async, opts: opts));
   }
 
@@ -645,8 +632,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/setbit
-  Pipeline setbit(String key, int offset, int bit,
-      [CommandOption<int, int>? opts]) {
+  Pipeline setbit(String key, int offset, int bit, [CommandOption<int, int>? opts]) {
     return _chain(SetBitCommand(key, offset, bit, opts));
   }
 
@@ -670,8 +656,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/setrange
-  Pipeline setrange(String key, int offset, String value,
-      [CommandOption<int, int>? opts]) {
+  Pipeline setrange(String key, int offset, String value, [CommandOption<int, int>? opts]) {
     return _chain(SetRangeCommand(key, offset, value, opts));
   }
 
@@ -684,14 +669,12 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/sinterstore
-  Pipeline sinterstore(String destination, List<String> keys,
-      [CommandOption<int, int>? opts]) {
+  Pipeline sinterstore(String destination, List<String> keys, [CommandOption<int, int>? opts]) {
     return _chain(SInterStoreCommand(destination, keys, opts));
   }
 
   /// @see https://redis.io/commands/sismember
-  Pipeline sismember<TData>(String key, TData member,
-      [CommandOption<dynamic, int>? opts]) {
+  Pipeline sismember<TData>(String key, TData member, [CommandOption<dynamic, int>? opts]) {
     return _chain(SIsMemberCommand<TData>(key, member, opts));
   }
 
@@ -732,8 +715,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/srem
-  Pipeline srem<TData>(String key, List<TData> members,
-      [CommandOption<int, int>? opts]) {
+  Pipeline srem<TData>(String key, List<TData> members, [CommandOption<int, int>? opts]) {
     return _chain(SRemCommand<TData>(key, members, opts));
   }
 
@@ -745,8 +727,7 @@ class Pipeline {
     int? count,
     CommandOption<List<dynamic>, List<dynamic>>? opts,
   }) {
-    return _chain(
-        SScanCommand(key, cursor, match: match, count: count, opts: opts));
+    return _chain(SScanCommand(key, cursor, match: match, count: count, opts: opts));
   }
 
   /// @see https://redis.io/commands/strlen
@@ -763,8 +744,7 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/sunionstore
-  Pipeline sunionstore(String destination, List<String> keys,
-      [CommandOption<int, int>? opts]) {
+  Pipeline sunionstore(String destination, List<String> keys, [CommandOption<int, int>? opts]) {
     return _chain(SUnionStoreCommand(destination, keys, opts));
   }
 
@@ -933,38 +913,32 @@ class Pipeline {
   }
 
   /// @see https://redis.io/commands/zrank
-  Pipeline zrank<TData>(String key, TData member,
-      [CommandOption<int?, int?>? opts]) {
+  Pipeline zrank<TData>(String key, TData member, [CommandOption<int?, int?>? opts]) {
     return _chain(ZRankCommand<TData>(key, member, opts));
   }
 
   /// @see https://redis.io/commands/zrem
-  Pipeline zrem<TData>(String key, List<TData> members,
-      [CommandOption<int, int>? opts]) {
+  Pipeline zrem<TData>(String key, List<TData> members, [CommandOption<int, int>? opts]) {
     return _chain(ZRemCommand<TData>(key, members, opts));
   }
 
   /// @see https://redis.io/commands/zremrangebylex
-  Pipeline zremrangebylex(String key, String min, String max,
-      [CommandOption<int, int>? opts]) {
+  Pipeline zremrangebylex(String key, String min, String max, [CommandOption<int, int>? opts]) {
     return _chain(ZRemRangeByLexCommand(key, min, max, opts));
   }
 
   /// @see https://redis.io/commands/zremrangebyrank
-  Pipeline zremrangebyrank(String key, int start, int stop,
-      [CommandOption<int, int>? opts]) {
+  Pipeline zremrangebyrank(String key, int start, int stop, [CommandOption<int, int>? opts]) {
     return _chain(ZRemRangeByRankCommand(key, start, stop, opts));
   }
 
   /// @see https://redis.io/commands/zremrangebyscore
-  Pipeline zremrangebyscore(String key, int min, int max,
-      [CommandOption<int, int>? opts]) {
+  Pipeline zremrangebyscore(String key, int min, int max, [CommandOption<int, int>? opts]) {
     return _chain(ZRemRangeByScoreCommand(key, min, max, opts));
   }
 
   /// @see https://redis.io/commands/zrevrank
-  Pipeline zrevrank<TData>(String key, TData member,
-      [CommandOption<int?, int?>? opts]) {
+  Pipeline zrevrank<TData>(String key, TData member, [CommandOption<int?, int?>? opts]) {
     return _chain(ZRevRankCommand<TData>(key, member, opts));
   }
 
@@ -976,13 +950,11 @@ class Pipeline {
     int? count,
     CommandOption<List<dynamic>, List<dynamic>>? opts,
   }) {
-    return _chain(
-        ZScanCommand(key, cursor, match: match, count: count, opts: opts));
+    return _chain(ZScanCommand(key, cursor, match: match, count: count, opts: opts));
   }
 
   /// @see https://redis.io/commands/zscore
-  Pipeline zscore<TData>(String key, TData member,
-      [CommandOption<String, num>? opts]) {
+  Pipeline zscore<TData>(String key, TData member, [CommandOption<String, num>? opts]) {
     return _chain(ZScoreCommand<TData>(key, member, opts));
   }